+++ gcovr/html_generator.py
@@ -8,23 +8,21 @@
 
 import os
 import sys
-import time
 import datetime
-import zlib
+import hashlib
 import io
 
 from .version import __version__
-from .utils import commonpath, sort_coverage
-from .coverage import FileCoverage
+from .utils import commonpath, sort_coverage, calculate_coverage, Logger
 
 
-class lazy(object):
+class Lazy:
     def __init__(self, fn):
 
-        def load():
-            result = fn()
+        def load(*args):
+            result = fn(*args)
 
-            def reuse_value():
+            def reuse_value(*args):
                 return result
 
             self.get = reuse_value
@@ -32,112 +30,276 @@ class lazy(object):
 
         self.get = load
 
-    def __call__(self):
-        return self.get()
+    def __call__(self, *args):
+        return self.get(*args)
 
 
 # Loading Jinja and preparing the environmen is fairly costly.
 # Only do this work if templates are actually used.
 # This speeds up text and XML output.
-@lazy
+@Lazy
 def templates():
     from jinja2 import Environment, PackageLoader
     return Environment(
         loader=PackageLoader('gcovr'),
-        autoescape=False,
+        autoescape=True,
         trim_blocks=True,
         lstrip_blocks=True)
 
 
-low_color = "LightPink"
-medium_color = "#FFFF55"
-high_color = "LightGreen"
-covered_color = "LightGreen"
-uncovered_color = "LightPink"
-takenBranch_color = "Green"
-notTakenBranch_color = "Red"
+@Lazy
+def user_templates():
+    from jinja2 import Environment, FunctionLoader
+
+    def load_user_template(template):
+        contents = None
+        try:
+            with open(template, "rb") as f:
+                contents = f.read().decode("utf-8")
+        # This exception can only occure if the file gets inaccesable while gcovr is running.
+        except Exception:  # pragma: no cover
+            pass
+
+        return contents
+
+    return Environment(
+        loader=FunctionLoader(load_user_template),
+        autoescape=True,
+        trim_blocks=True,
+        lstrip_blocks=True)
 
 
-def html_escape(s):
-    """Escape string for inclusion in a HTML body.
+class CssRenderer():
 
-    Does not escape ``'``, ``"``, or ``>``.
-    """
-    s = s.replace('&', '&amp;')
-    s = s.replace('<', '&lt;')
-    return s
+    Themes = {
+        'green': {
+            'unknown_color': "LightGray",
+            'low_color': "#FF6666",
+            'medium_color': "#F9FD63",
+            'high_color': "#85E485",
+            'covered_color': "#85E485",
+            'uncovered_color': "#FF8C8C",
+            'takenBranch_color': "Green",
+            'notTakenBranch_color': "Red"
+        },
+        'blue': {
+            'unknown_color': "LightGray",
+            'low_color': "#FF6666",
+            'medium_color': "#F9FD63",
+            'high_color': "#66B4FF",
+            'covered_color': "#66B4FF",
+            'uncovered_color': "#FF8C8C",
+            'takenBranch_color': "Blue",
+            'notTakenBranch_color': "Red"
+        }
+    }
+
+    @staticmethod
+    def get_themes():
+        return list(CssRenderer.Themes.keys())
+
+    @staticmethod
+    def get_default_theme():
+        return 'green'
+
+    @staticmethod
+    def render(options):
+        template = None
+        if options.html_css is not None:
+            template = user_templates().get_template(os.path.relpath(options.html_css))
+        else:
+            template = templates().get_template('style.css')
+        return template.render(
+            CssRenderer.Themes[options.html_theme],
+            tab_size=options.html_tab_size
+        )
 
 
-def calculate_coverage(covered, total, nan_value=0.0):
-    return nan_value if total == 0 else round(100.0 * covered / total, 1)
+class NullHighlighting:
+    def get_css(self):
+        return ''
+
+    @staticmethod
+    def highlighter_for_file(filename):
+        return lambda code: [line.rstrip() for line in code.split("\n")]
 
 
-def coverage_to_color(coverage, medium_threshold, high_threshold):
+class PygmentHighlighting:
+    def __init__(self, options):
+        self.logger = Logger(options.verbose)
+        self.formatter = None
+        try:
+            from pygments.formatters.html import HtmlFormatter
+            self.formatter = HtmlFormatter(nowrap=True)
+        except ImportError as e:  # pragma: no cover
+            self.logger.warn("No syntax highlighting availabel: {}".format(str(e)))
+
+    def get_css(self):
+        if self.formatter is None:  # pragma: no cover
+            return ''
+        return "\n\n/* pygments syntax highlighting */\n" + self.formatter.get_style_defs()
+
+    # Set the lexer for the given filename. Return true if a lexer is found
+    def highlighter_for_file(self, filename):
+        if self.formatter is None:  # pragma: no cover
+            return NullHighlighting.highlighter_for_file(filename)
+
+        import pygments
+        from pygments.lexers import get_lexer_for_filename
+        from jinja2 import Markup
+        try:
+            lexer = get_lexer_for_filename(filename, None, stripnl=False)
+            return lambda code: [Markup(line.rstrip()) for line in pygments.highlight(code, lexer, self.formatter).split("\n")]
+        except pygments.util.ClassNotFound:  # pragma: no cover
+            return NullHighlighting.highlighter_for_file(filename)
+
+
+@Lazy
+def get_formatter(options):
+    return PygmentHighlighting(options) if options.html_details_syntax_highlighting else NullHighlighting()
+
+
+def coverage_to_class(coverage, medium_threshold, high_threshold):
     if coverage is None:
-        return 'LightGray'
-    elif coverage < medium_threshold:
-        return low_color
-    elif coverage < high_threshold:
-        return medium_color
-    else:
-        return high_color
+        return 'coverage-unknown'
+    if coverage == 0:
+        return 'coverage-none'
+    if coverage < medium_threshold:
+        return 'coverage-low'
+    if coverage < high_threshold:
+        return 'coverage-medium'
+    return 'coverage-high'
+
+
+class RootInfo:
+
+    def __init__(self, options):
+        self.medium_threshold = options.html_medium_threshold
+        self.high_threshold = options.html_high_threshold
+        self.details = options.html_details
+        self.relative_anchors = options.relative_anchors
+
+        self.version = __version__
+        self.head = options.html_title
+        self.date = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+        self.encoding = options.html_encoding
+        self.directory = None
+        self.branches = dict()
+        self.lines = dict()
+        self.files = []
+
+    def set_directory(self, directory):
+        self.directory = directory
+
+    def get_directory(self):
+        return "." if self.directory == '' else self.directory.replace('\\', '/')
+
+    def calculate_branch_coverage(self, covdata):
+        branch_total = 0
+        branch_covered = 0
+        for key in covdata.keys():
+            (total, covered, _percent) = covdata[key].branch_coverage()
+            branch_total += total
+            branch_covered += covered
+        self.branches['exec'] = branch_covered
+        self.branches['total'] = branch_total
+        coverage = calculate_coverage(branch_covered, branch_total, nan_value=None)
+        self.branches['coverage'] = '-' if coverage is None else coverage
+        self.branches['class'] = self._coverage_to_class(coverage)
+
+    def calculate_line_coverage(self, covdata):
+        line_total = 0
+        line_covered = 0
+        for key in covdata.keys():
+            (total, covered, _percent) = covdata[key].line_coverage()
+            line_total += total
+            line_covered += covered
+        self.lines['exec'] = line_covered
+        self.lines['total'] = line_total
+        coverage = calculate_coverage(line_covered, line_total)
+        self.lines['coverage'] = coverage
+        self.lines['class'] = self._coverage_to_class(coverage)
+
+    def add_file(self, cdata, link_report, cdata_fname):
+        lines_total, lines_exec, _ = cdata.line_coverage()
+        branches_total, branches_exec, _ = cdata.branch_coverage()
+
+        line_coverage = calculate_coverage(
+            lines_exec, lines_total, nan_value=100.0)
+        branch_coverage = calculate_coverage(
+            branches_exec, branches_total, nan_value=None)
+
+        lines = {
+            'total': lines_total,
+            'exec': lines_exec,
+            'coverage': line_coverage,
+            'class': self._coverage_to_class(line_coverage),
+        }
+
+        branches = {
+            'total': branches_total,
+            'exec': branches_exec,
+            'coverage': '-' if branch_coverage is None else branch_coverage,
+            'class': self._coverage_to_class(branch_coverage),
+        }
+
+        display_filename = (
+            os.path.relpath(os.path.realpath(cdata_fname), self.directory)
+            .replace('\\', '/'))
+
+        if link_report is not None:
+            if self.relative_anchors:
+                link_report = os.path.basename(link_report)
+
+        self.files.append(dict(
+            directory=self.directory,
+            filename=display_filename,
+            link=link_report,
+            lines=lines,
+            branches=branches,
+        ))
+
+    def _coverage_to_class(self, coverage):
+        return coverage_to_class(coverage, self.medium_threshold, self.high_threshold)
 
 
 #
 # Produce an HTML report
 #
 def print_html_report(covdata, output_file, options):
+    css_data = CssRenderer.render(options)
     medium_threshold = options.html_medium_threshold
     high_threshold = options.html_high_threshold
-    details = options.html_details
-    if output_file is None:
-        details = False
+
     data = {}
-    data['HEAD'] = options.html_title
-    data['VERSION'] = __version__
-    data['TIME'] = str(int(time.time()))
-    data['DATE'] = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-    data['ROWS'] = []
-    data['ENC'] = options.html_encoding
-    data['low_color'] = low_color
-    data['medium_color'] = medium_color
-    data['high_color'] = high_color
+    root_info = RootInfo(options)
+    data['info'] = root_info
+
     data['COVERAGE_MED'] = medium_threshold
     data['COVERAGE_HIGH'] = high_threshold
-    data['CSS'] = templates().get_template('style.css').render(
-        low_color=low_color,
-        medium_color=medium_color,
-        high_color=high_color,
-        covered_color=covered_color,
-        uncovered_color=uncovered_color,
-        takenBranch_color=takenBranch_color,
-        notTakenBranch_color=notTakenBranch_color
-    )
-    data['DIRECTORY'] = ''
 
-    branchTotal = 0
-    branchCovered = 0
-    for key in covdata.keys():
-        (total, covered, percent) = covdata[key].branch_coverage()
-        branchTotal += total
-        branchCovered += covered
-    data['BRANCHES_EXEC'] = str(branchCovered)
-    data['BRANCHES_TOTAL'] = str(branchTotal)
-    coverage = calculate_coverage(branchCovered, branchTotal, nan_value=None)
-    data['BRANCHES_COVERAGE'] = '-' if coverage is None else str(coverage)
-    data['BRANCHES_COLOR'] = coverage_to_color(coverage, medium_threshold, high_threshold)
+    self_contained = options.html_self_contained
+    if self_contained is None:
+        self_contained = not options.html_details
 
-    lineTotal = 0
-    lineCovered = 0
-    for key in covdata.keys():
-        (total, covered, percent) = covdata[key].line_coverage()
-        lineTotal += total
-        lineCovered += covered
-    data['LINES_EXEC'] = str(lineCovered)
-    data['LINES_TOTAL'] = str(lineTotal)
-    coverage = calculate_coverage(lineCovered, lineTotal)
-    data['LINES_COVERAGE'] = str(coverage)
-    data['LINES_COLOR'] = coverage_to_color(coverage, medium_threshold, high_threshold)
+    formatter = get_formatter(options)
+    css_data += formatter.get_css()
+
+    if self_contained:
+        data['css'] = css_data
+    else:
+        css_output = os.path.splitext(output_file)[0] + '.css'
+        with open(css_output, 'w') as f:
+            f.write(css_data)
+
+        if options.relative_anchors:
+            css_link = os.path.basename(css_output)
+        else:
+            css_link = css_output
+        data['css_link'] = css_link
+
+    root_info.calculate_branch_coverage(covdata)
+    root_info.calculate_line_coverage(covdata)
 
     # Generate the coverage output (on a per-package basis)
     # source_dirs = set()
@@ -156,69 +318,40 @@ def print_html_report(covdata, output_file, options):
         files.append(filtered_fname)
         dirs.append(os.path.dirname(filtered_fname) + os.sep)
         cdata_fname[f] = filtered_fname
-        if not details:
-            cdata_sourcefile[f] = None
-        else:
+        if options.html_details:
             cdata_sourcefile[f] = _make_short_sourcename(
                 output_file, filtered_fname)
+        else:
+            cdata_sourcefile[f] = None
 
     # Define the common root directory, which may differ from options.root
     # when source files share a common prefix.
+    root_directory = ''
     if len(files) > 1:
         commondir = commonpath(files)
         if commondir != '':
-            data['DIRECTORY'] = commondir
+            root_directory = commondir
     else:
-        dir_, file_ = os.path.split(filtered_fname)
+        dir_, _file = os.path.split(filtered_fname)
         if dir_ != '':
-            data['DIRECTORY'] = dir_ + os.sep
+            root_directory = dir_ + os.sep
 
-    nrows = 0
+    root_info.set_directory(root_directory)
+
     for f in keys:
-        cdata = covdata[f]  # type: FileCoverage
-        class_lines = 0
-        class_hits = 0
-        class_branches = 0
-        class_branch_hits = 0
-        class_lines, class_hits, _ = cdata.line_coverage()
-        b_total, b_hits, _ = cdata.branch_coverage()
-        class_branch_hits += b_hits
-        class_branches += b_total
+        root_info.add_file(covdata[f], cdata_sourcefile[f], cdata_fname[f])
 
-        lines_covered = calculate_coverage(
-            class_hits, class_lines, nan_value=100.0)
-        branches_covered = calculate_coverage(
-            class_branch_hits, class_branches, nan_value=None)
-
-        nrows += 1
-        data['ROWS'].append(html_row(
-            options, details, cdata_sourcefile[f], nrows,
-            directory=data['DIRECTORY'],
-            filename=os.path.relpath(
-                os.path.realpath(cdata_fname[f]), data['DIRECTORY']),
-            LinesExec=class_hits,
-            LinesTotal=class_lines,
-            LinesCoverage=lines_covered,
-            BranchesExec=class_branch_hits,
-            BranchesTotal=class_branches,
-            BranchesCoverage=branches_covered
-        ))
-
-    if data['DIRECTORY'] == '':
-        data['DIRECTORY'] = "."
-    data['DIRECTORY'] = data['DIRECTORY'].replace('\\', '/')
-
-    htmlString = templates().get_template('root_page.html').render(**data)
+    html_string = templates().get_template('root_page.html').render(**data)
 
     if output_file is None:
-        sys.stdout.write(htmlString + '\n')
+        sys.stdout.write(html_string + '\n')
     else:
         with io.open(output_file, 'w', encoding=options.html_encoding,
                      errors='xmlcharrefreplace') as fh:
-            fh.write(htmlString + '\n')
+            fh.write(html_string + '\n')
 
     # Return, if no details are requested
-    if not details:
+    if not options.html_details:
         return
 
     #
@@ -227,110 +360,82 @@ def print_html_report(covdata, output_file, options):
     for f in keys:
         cdata = covdata[f]
 
-        data['FILENAME'] = cdata_fname[f]
-        data['ROWS'] = ''
+        data['filename'] = cdata_fname[f]
 
-        branchTotal, branchCovered, coverage = cdata.branch_coverage()
-        data['BRANCHES_EXEC'] = str(branchCovered)
-        data['BRANCHES_TOTAL'] = str(branchTotal)
-        data['BRANCHES_COVERAGE'] = '-' if coverage is None else str(coverage)
-        data['BRANCHES_COLOR'] = coverage_to_color(coverage, medium_threshold, high_threshold)
+        branches = dict()
+        data['branches'] = branches
 
-        lineTotal, lineCovered, coverage = cdata.line_coverage()
-        data['LINES_EXEC'] = str(lineCovered)
-        data['LINES_TOTAL'] = str(lineTotal)
-        data['LINES_COVERAGE'] = str(coverage)
-        data['LINES_COLOR'] = coverage_to_color(coverage, medium_threshold, high_threshold)
+        branches['total'], branches['exec'], branches['coverage'] = cdata.branch_coverage()
+        branches['class'] = coverage_to_class(branches['coverage'], medium_threshold, high_threshold)
+        branches['coverage'] = '-' if branches['coverage'] is None else branches['coverage']
 
-        data['ROWS'] = []
+        lines = dict()
+        data['lines'] = lines
+        lines['total'], lines['exec'], lines['coverage'] = cdata.line_coverage()
+        lines['class'] = coverage_to_class(lines['coverage'], medium_threshold, high_threshold)
+
+        data['source_lines'] = []
         currdir = os.getcwd()
         os.chdir(options.root_dir)
-        with io.open(data['FILENAME'], 'r', encoding=options.source_encoding,
-                     errors='replace') as INPUT:
-            for ctr, line in enumerate(INPUT, 1):
-                data['ROWS'].append(
-                    source_row(ctr, line.rstrip(), cdata.lines.get(ctr))
+        with io.open(data['filename'], 'r', encoding=options.source_encoding,
+                     errors='replace') as source_file:
+            lines = formatter.highlighter_for_file(data['filename'])(source_file.read())
+            for ctr, line in enumerate(lines, 1):
+                data['source_lines'].append(
+                    source_row(ctr, line, cdata.lines.get(ctr))
                 )
         os.chdir(currdir)
 
-        htmlString = templates().get_template('source_page.html').render(**data)
+        html_string = templates().get_template('source_page.html').render(**data)
         with io.open(cdata_sourcefile[f], 'w', encoding=options.html_encoding,
                      errors='xmlcharrefreplace') as fh:
-            fh.write(htmlString + '\n')
+            fh.write(html_string + '\n')
 
 
 def source_row(lineno, source, line_cov):
-    kwargs = {}
-    kwargs['lineno'] = str(lineno)
-    kwargs['linebranch'] = []
-    if line_cov and line_cov.is_covered:
-        kwargs['covclass'] = 'coveredLine'
-        # If line has branches them show them with ticks or crosses
-        branches = line_cov.branches
-        branchcounter = 0
-        for branch_id in sorted(branches):
-            branchcounter += 1
-            branch = branches[branch_id]
-            branch_args = {}
-            if branch.is_covered:
-                branch_args['class'] = 'takenBranch'
-                branch_args['message'] = 'Branch {name} taken {count} times'.format(
-                    name=branch_id, count=branch.count)
-                branch_args['symbol'] = '&check;'
-            else:
-                branch_args['class'] = 'notTakenBranch'
-                branch_args['message'] = 'Branch {name} not taken'.format(
-                    name=branch_id)
-                branch_args['symbol'] = '&cross;'
-            branch_args['wrap'] = (branchcounter % 4) == 0
-            kwargs['linebranch'].append(branch_args)
-        kwargs['linecount'] = str(line_cov.count)
-    elif line_cov and line_cov.is_uncovered:
-        kwargs['covclass'] = 'uncoveredLine'
-        kwargs['linebranch'] = ''
-        kwargs['linecount'] = ''
-    else:
-        kwargs['covclass'] = ''
-        kwargs['linebranch'] = ''
-        kwargs['linecount'] = ''
-    kwargs['source'] = html_escape(source)
-    return kwargs
+    linebranch = None
+    linecount = ''
+    covclass = ''
+    if line_cov:
+        if line_cov.is_covered:
+            covclass = 'coveredLine'
+            linebranch = source_row_branch(line_cov.branches)
+            linecount = line_cov.count
+        elif line_cov.is_uncovered:
+            covclass = 'uncoveredLine'
+    return {
+        'lineno': lineno,
+        'source': source,
+        'covclass': covclass,
+        'linebranch': linebranch,
+        'linecount': linecount,
+    }
 
 
-#
-# Generate the table row for a single file
-#
-def html_row(options, details, sourcefile, nrows, **kwargs):
-    if details and options.relative_anchors:
-        sourcefile = os.path.basename(sourcefile)
-    if nrows % 2 == 0:
-        kwargs['altstyle'] = 'style="background-color:LightSteelBlue"'
-    else:
-        kwargs['altstyle'] = ''
-    if details:
-        kwargs['filename'] = '<a href="%s">%s</a>' % (
-            sourcefile, kwargs['filename'].replace('\\', '/')
-        )
-    kwargs['LinesCoverage'] = round(kwargs['LinesCoverage'], 1)
-    # Disable the border if the bar is too short to see the color
-    if kwargs['LinesCoverage'] < 1e-7:
-        kwargs['BarBorder'] = "border:white; "
-    else:
-        kwargs['BarBorder'] = ""
-    if kwargs['LinesCoverage'] < options.html_medium_threshold:
-        kwargs['LinesColor'] = low_color
-        kwargs['LinesBar'] = 'red'
-    elif kwargs['LinesCoverage'] < options.html_high_threshold:
-        kwargs['LinesColor'] = medium_color
-        kwargs['LinesBar'] = 'yellow'
-    else:
-        kwargs['LinesColor'] = high_color
-        kwargs['LinesBar'] = 'green'
+def source_row_branch(branches):
+    if not branches:
+        return None
 
-    kwargs['BranchesColor'] = coverage_to_color(kwargs['BranchesCoverage'], options.html_medium_threshold, options.html_high_threshold)
-    kwargs['BranchesCoverage'] = '-' if kwargs['BranchesCoverage'] is None else round(kwargs['BranchesCoverage'], 1)
+    taken = 0
+    total = 0
+    items = []
 
-    return kwargs
+    for branch_id in sorted(branches):
+        branch = branches[branch_id]
+        if branch.is_covered:
+            taken += 1
+        total += 1
+        items.append({
+            'taken': branch.is_covered,
+            'name': branch_id,
+            'count': branch.count,
+        })
+
+    return {
+        'taken': taken,
+        'total': total,
+        'branches': items,
+    }
 
 
 def _make_short_sourcename(output_file, filename):
@@ -342,23 +447,12 @@ def _make_short_sourcename(output_file, filename):
         filename (str): Path from root to source code.
     """
 
-    output_file_parts = os.path.abspath(output_file).split('.')
-    if len(output_file_parts) > 1:
-        output_prefix = '.'.join(output_file_parts[:-1])
-        output_suffix = output_file_parts[-1]
-    else:
-        output_prefix = output_file
-        output_suffix = 'html'
+    (output_prefix, output_suffix) = os.path.splitext(os.path.abspath(output_file))
+    if output_suffix == '':
+        output_suffix = '.html'
 
-    longname = filename.replace(os.sep, '_')
-    longname_hash = ""
-    while True:
-        sourcename = '.'.join((
-            output_prefix, longname + longname_hash, output_suffix))
-        # we add a hash at the end and attempt to shorten the
-        # filename if it exceeds common filesystem limitations
-        if len(os.path.basename(sourcename)) < 256:
-            break
-        longname_hash = "_" + hex(zlib.crc32(longname) & 0xffffffff)[2:]
-        longname = longname[(len(sourcename) - len(longname_hash)):]
+    filename = filename.replace(os.sep, '/')
+    sourcename = '.'.join((output_prefix,
+                          os.path.basename(filename),
+                          hashlib.md5(filename.encode('utf-8')).hexdigest())) + output_suffix
     return sourcename